fix： CSV NULL/空串；无过滤条件的 filter 导出，inspect_export 直接用 Parquet 元数据行数（仍校验 fingerprint/编译查询），不再执行 COUNT；SQL 导出因行数无法从元数据得知，保留 COUNT；后端新增 AppError::AlreadyExists → code ALREADY_EXISTS，前端 App.tsx 改为按 code 判断，去掉 /already exists/i 正则

+++ src-tauri/src/query/worker.rs
@@ -354,7 +354,9 @@ fn available_space(path: &Path) -> Option<u64> {
         return None;
     }
     let stats = unsafe { stats.assume_init() };
-    Some((stats.f_bavail as u64).saturating_mul(stats.f_frsize))
+    // The cast is required on platforms where `f_bavail` is not already `u64`.
+    #[allow(clippy::unnecessary_cast)]
+    Some((stats.f_bavail as u64).saturating_mul(stats.f_frsize as u64))
 }
 
 #[cfg(not(unix))]
